minor refactoring in some files

+++ src/game.rs
@@ -16,7 +16,7 @@ impl Game {
             current_piece: Some(Piece::from(random::<PieceType>())),
             score: 0,
             lines_cleared: 0,
-            game_map: vec![vec![Presence::No; NUM_BLOCKS_X as usize]; NUM_BLOCKS_Y as usize],
+            game_map: vec![vec![Presence::No; NUM_BLOCKS_X]; NUM_BLOCKS_Y],
         }
     }
 }
